reinstate multi input count and add tests
