+++ spark/src/main/scala/org/apache/comet/serde/aggregates.scala
@@ -109,10 +109,6 @@ object CometCount extends CometAggregateExpressionSerde[Count] {
       inputs: Seq[Attribute],
       binding: Boolean,
       conf: SQLConf): Option[ExprOuterClass.AggExpr] = {
-    if (expr.children.length > 1) {
-      withInfo(expr, "Count only supported with a single argument")
-      return None
-    }
     val exprChildren = expr.children.map(exprToProto(_, inputs, binding))
     if (exprChildren.forall(_.isDefined)) {
       val builder = ExprOuterClass.Count.newBuilder()
